Trait esercizio5::stati::stati::State

pub trait State {
    // Required methods
    fn deposita(&self, saldo: &mut f64, amount: f64);
    fn preleva(&self, saldo: &mut f64, amount: f64);
    fn paga_interessi(&self, saldo: &mut f64, interesse: f64);
    fn stato(&self) -> Stato;
}

Definisce il comportamento di uno stato del conto bancario.

Required Methods

fn deposita(&self, saldo: &mut f64, amount: f64)

Deposita una certa quantità di denaro nel conto.

fn preleva(&self, saldo: &mut f64, amount: f64)

Preleva una certa quantità di denaro dal conto.

fn paga_interessi(&self, saldo: &mut f64, interesse: f64)

Calcola e aggiunge gli interessi al saldo del conto.

fn stato(&self) -> Stato

Restituisce lo stato attuale del conto.